Some changes to fix compiler warnings. Issue #8063.
git-svn-id: http://svn.code.sf.net/p/postbooks/code/xtuple/trunk@4686 ea1a15ef-c238-0410-93cb-fa28e9f07733

+++ guiclient/salesOrder.cpp
@@ -211,7 +211,7 @@ salesOrder::salesOrder(QWidget* parent, const char* name, Qt::WFlags fl)
   _soitem->addColumn(tr("Firm"),                 0, Qt::AlignCenter,false, "coitem_firm");
   _soitem->addColumn(tr("Sched. Date"),_dateColumn, Qt::AlignCenter,true, "coitem_scheddate");
   _soitem->addColumn(tr("Ordered"),     _qtyColumn, Qt::AlignRight, true, "coitem_qtyord");
-  _soitem->addColumn(tr("Qty UOM"),     _uomColumn*1.5, Qt::AlignLeft,  true, "qty_uom");
+  _soitem->addColumn(tr("Qty UOM"),     (int)(_uomColumn*1.5), Qt::AlignLeft,  true, "qty_uom");
   _soitem->addColumn(tr("Shipped"),     _qtyColumn, Qt::AlignRight, true, "qtyshipped");
   _soitem->addColumn(tr("At Shipping"), _qtyColumn, Qt::AlignRight, false, "qtyatshipping");
   _soitem->addColumn(tr("Balance"),     _qtyColumn, Qt::AlignRight, false, "balance");
